# BUG: sparse: Remove reference cycle to improve memory use