@andyfaff 2025-03-30
I'm working on the `VectorFunction` reference cycle.