@andyfaff 2025-03-31
This looks fine to me. Regarding the extra overhead, how many times would people want to do the round trip on the MLO (o.adjoint().adjoint())?
I'll leave it up to some linalg people to merge though.